refactor: split ssh file operations api
